consolidate blacklist + forbidden
